+++ src/sess.rs
@@ -9,6 +9,7 @@ use std;
 use std::collections::{HashMap, HashSet};
 use std::fmt;
 use std::fs::canonicalize;
+use std::io::Write;
 use std::iter::FromIterator;
 use std::mem::swap;
 use std::path::{Path, PathBuf};
@@ -828,6 +829,54 @@ impl<'io, 'sess: 'io, 'ctx: 'sess> SessionIo<'sess, 'ctx> {
         Ok(path)
     }
 
+    /// Checkout only git dependency's path sub-dependency Bender.yml files
+    async fn sub_dependency_fixing(
+        &'io self,
+        dep_iter_mut: &mut HashMap<String, config::Dependency>,
+        top_package_name: String,
+        top_dep_id: DependencyRef,
+        db: Git<'sess, 'ctx>,
+        used_git_rev: &str,
+    ) -> Result<()> {
+        for dep in (dep_iter_mut).iter_mut() {
+            if let (_, config::Dependency::Path(ref path)) = dep {
+                if !path.starts_with("/") {
+                    warnln!("Path dependencies ({:?}) in git dependencies ({:?}) currently not fully supported. Your mileage may vary.", dep.0, top_package_name);
+
+                    let sub_entries = db
+                        .list_files(used_git_rev, Some(path.join("Bender.yml")))
+                        .await?;
+                    let sub_data = match sub_entries.into_iter().next() {
+                        None => Ok(None),
+                        Some(sub_entry) => db.cat_file(sub_entry.hash).await.map(Some),
+                    }?;
+
+                    let sub_dep_path = self.get_package_path(top_dep_id).join(path).clone();
+
+                    if let Some(full_sub_data) = sub_data {
+                        if !sub_dep_path.exists() {
+                            std::fs::create_dir_all(sub_dep_path.clone())?;
+                        }
+                        let mut sub_file = std::fs::OpenOptions::new()
+                            .write(true)
+                            .truncate(true)
+                            .create(true)
+                            .open(sub_dep_path.clone().join("Bender.yml"))?;
+                        writeln!(&mut sub_file, "{}", full_sub_data)?;
+                        sub_file.flush()?;
+                    }
+
+                    *dep.1 = config::Dependency::Path(
+                        self.get_package_path(top_dep_id).join(path).clone(),
+                    );
+
+                    // TODO: fix possible recursion for further path dependencies
+                }
+            }
+        }
+        Ok(())
+    }
+
     /// Load the manifest for a specific version of a dependency.
     ///
     /// Loads and returns the manifest for a dependency at a specific version.
@@ -875,6 +924,7 @@ impl<'io, 'sess: 'io, 'ctx: 'sess> SessionIo<'sess, 'ctx> {
                         Err(e) => Err(e),
                     }
                 } else {
+                    warnln!("Manifest not found for {:?}", dep.name);
                     Ok(None)
                 }
             }
@@ -913,22 +963,23 @@ impl<'io, 'sess: 'io, 'ctx: 'sess> SessionIo<'sess, 'ctx> {
                                 cause,
                             )
                         })?;
+
                         // Add base path to path dependencies within git repositories
-                        for dep in full.dependencies.iter_mut() {
-                            if let (_, config::Dependency::Path(ref path)) = dep {
-                                if !path.starts_with("/") {
-                                    if !self.get_package_path(dep_id).exists() {
-                                        warnln!("Please note that dependencies for {:?} may not be available unless {:?} is properly checked out.\n         (to checkout run `bender sources` and then `bender update` again).", dep.0, full.package.name);
-                                    }
-                                    *dep.1 = config::Dependency::Path(
-                                        self.get_package_path(dep_id).join(path).clone(),
-                                    );
-                                }
-                            }
-                        }
+                        self.sub_dependency_fixing(
+                            &mut full.dependencies,
+                            full.package.name.clone(),
+                            dep_id,
+                            db,
+                            rev,
+                        )
+                        .await?;
+
                         Ok(Some(self.sess.intern_manifest(full)))
                     }
-                    None => Ok(None),
+                    None => {
+                        warnln!("Manifest not found for {:?}", dep.name);
+                        Ok(None)
+                    }
                 };
                 let manifest = manifest?;
                 self.sess