+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Changed
 - Update `tokio` dependency, update to `futures=0.3`, rewrite for `async/await`. Bumps minimum rust to 1.57
 - Update `serde_yaml` dependency to `0.9`. Bumps minimum rust to 1.58
+- Partially checkout path dependencies in git dependencies
+- Add warnings if manifest file is not found for a dependency
 
 ## 0.25.3 - 2022-08-05
 ### Added